add styles.css, add link to .html files
Still learning git, html, and now CSS. I added a styles.css file, and
added the links to the the styles.css in the .html files. of note I
need to remember that when linking files that I take into account the
directory structure (i.e. if I want to link to a file in a parent
directory I need to use ../filename.html).
Also, I made these changes on the sandbox-odin-reciepies-July-30-2023
branch